+++ lib/hashed_kv_store.dart
@@ -1,2 +1,3 @@
 export 'hashed_kv_path.dart';
+export 'kv_layout_version.dart';
 export 'multi_isolate_kv_store_client.dart';